#82 Prepare for versioned release
