+++ checkself/checkself_test.go
@@ -73,20 +73,6 @@ func TestBasicCheckSelf_CheckDatabaseSchema_badDatabaseURL(test *testing.T) {
 	assert.Equal(test, expected, newReportErrors[0])
 }
 
-func TestBasicCheckSelf_CheckDatabaseSchema_noSchemaInstalled(test *testing.T) {
-	ctx := context.TODO()
-	expected := "Senzing database schema has not been installed in sqlite3://na:na@/tmp/sqlite/G2C-empty.db. For more information, visit https://hub.senzing.com/...  Error: no such table: DSRC_RECORD"
-	testObject := getTestObject(ctx, test)
-	badReportErrors := []string{}
-	testObject.DatabaseURL = "sqlite3://na:na@/tmp/sqlite/G2C-empty.db"
-	newReportChecks, newReportInfo, newReportErrors, err := testObject.CheckDatabaseSchema(ctx, reportChecks(), reportInfo(), badReportErrors)
-	require.NoError(test, err)
-	assert.Len(test, newReportChecks, 1)
-	assert.Empty(test, newReportInfo)
-	assert.Len(test, newReportErrors, 1)
-	assert.Equal(test, expected, newReportErrors[0])
-}
-
 func TestBasicCheckSelf_CheckLicense(test *testing.T) {
 	ctx := context.TODO()
 	testObject := getTestObject(ctx, test)
@@ -118,32 +104,6 @@ func TestBasicCheckSelf_CheckLicense_badGetDatabaseURL(test *testing.T) {
 	assert.Len(test, newReportChecks, 1)
 	assert.Empty(test, newReportInfo)
 	assert.Len(test, newReportErrors, 1)
-}
-
-func TestBasicCheckSelf_CheckLicense_badGetLicense(test *testing.T) {
-	ctx := context.TODO()
-	expected := "Could not get count of records.  Error no such table: DSRC_RECORD"
-	testObject := getTestObject(ctx, test)
-	testObject.Settings = `
-        {
-            "PIPELINE": {
-                "CONFIGPATH": "/etc/opt/senzing",
-                "LICENSESTRINGBASE64": "badLicense",
-                "RESOURCEPATH": "/opt/senzing/g2/resources",
-                "SUPPORTPATH": "/opt/senzing/data"
-            },
-            "SQL": {
-                "BACKEND": "SQL",
-                "CONNECTION": "sqlite3://na:na@/tmp/sqlite/G2C-empty.db"
-            }
-        }
-        `
-	newReportChecks, newReportInfo, newReportErrors, err := testObject.CheckLicense(ctx, reportChecks(), reportInfo(), reportErrors())
-	require.NoError(test, err)
-	assert.Len(test, newReportChecks, 1)
-	assert.Empty(test, newReportInfo)
-	assert.Len(test, newReportErrors, 1)
-	assert.Equal(test, expected, newReportErrors[0])
 }
 
 func TestBasicCheckSelf_CheckSelf(test *testing.T) {
@@ -203,31 +163,6 @@ func TestBasicCheckSelf_CheckSettings_badSettings(test *testing.T) {
 	assert.Empty(test, newReportInfo)
 	assert.Len(test, newReportErrors, 1)
 	assert.Equal(test, expected, newReportErrors[0])
-}
-
-func TestBasicCheckSelf_CheckSettings_badDatabaseURLs(test *testing.T) {
-	ctx := context.TODO()
-	// expected := "????"
-	testObject := getTestObject(ctx, test)
-	testObject.Settings = `
-        {
-            "PIPELINE": {
-                "CONFIGPATH": "/etc/opt/senzing",
-                "LICENSESTRINGBASE64": "",
-                "RESOURCEPATH": "/opt/senzing/g2/resources",
-                "SUPPORTPATH": "/opt/senzing/data"
-            },
-            "SQL": {
-                "BACKEND": "SQL"
-            }
-        }
-        `
-	newReportChecks, newReportInfo, newReportErrors, err := testObject.CheckSettings(ctx, reportChecks(), reportInfo(), reportErrors())
-	require.NoError(test, err)
-	assert.Len(test, newReportChecks, 1)
-	assert.Empty(test, newReportInfo)
-	assert.Len(test, newReportErrors, 2)
-	// assert.Equal(test, expected, newReportErrors[0])
 }
 
 // ----------------------------------------------------------------------------
@@ -325,14 +260,6 @@ func TestBasicCheckSelf_checkDatabaseURL_badSqliteURL(test *testing.T) {
 	badDatabaseURL := "sqlite3://na:na@host.com:port//tmp/nodatabase.db"
 	actual := checkDatabaseURL(ctx, variableName, badDatabaseURL)
 	sink(expected, actual)
-}
-
-func TestBasicCheckSelf_checkDatabaseURL_badSqliteURL_stat(test *testing.T) {
-	ctx := context.TODO()
-	expected := "VariableName = sqlite3://na:na@/tmp/nodatabase.db is misconfigured. Could not find /tmp/nodatabase.db. For more information, visit https://hub.senzing.com/..."
-	badDatabaseURL := "sqlite3://na:na@/tmp/nodatabase.db"
-	actual := checkDatabaseURL(ctx, variableName, badDatabaseURL)
-	assert.Equal(test, expected, actual[0])
 }
 
 func TestBasicCheckSelf_checkDatabaseURL_badSchemaLength(test *testing.T) {
